requirements.txt
Django==1.8.10
django-haystack==2.4.1
whoosh==2.7.0
mistune==0.7.1
Pillow==3.1.0
django-infinite-scroll-pagination>=0.1.3,<0.2
unicode-slugify==0.1.3
django-allauth==0.26.1
django-spirit==0.4.7
django-djconfig==0.4.0
MySQL-python==1.2.3
sqlparse==0.1.19
oauth2==1.9.0.post1
oauthlib==1.0.3
requests==2.9.1
requests-oauthlib==0.6.2
gunicorn==19.4.5
gevent==1.0.1
greenlet==0.4.10
raven==5.13.0
django-qiniu-storage==2.3.0
qiniu==7.0.5
djangorestframework==3.1.3
redis-2.10.5
django-redis-cache==1.6.5


##################### Debug Requirement #########################
django_debug_toolbar==1.4



